+++ src/dataconnect/provisionCloudSql.ts
@@ -3,6 +3,7 @@ import * as utils from "../utils";
 import { grantRolesToCloudSqlServiceAccount } from "./checkIam";
 import { Instance } from "../gcp/cloudsql/types";
 import { promiseWithSpinner } from "../utils";
+import { logger } from "../logger";
 
 const GOOGLE_ML_INTEGRATION_ROLE = "roles/aiplatform.user";
 
@@ -106,6 +107,7 @@ export async function provisionCloudSql(args: {
     } else {
       // Skip it if the database is not accessible.
       // Possible that the CSQL instance is in the middle of something.
+      logger.debug(`Unexpected error from CloudSQL: ${err}`);
       silent || utils.logLabeledWarning("dataconnect", `Database ${databaseId} is not accessible.`);
     }
   }